Refs Task #397809 unit test show detail Admin Payload History
